// src/types/productType.ts
import { MyCartType } from "./cartType";
import { Categories } from "./categoryType";
import { Collection } from "./collectionType";
import { ProductStock } from "./productStock";
import { ProductRatingType } from "./reviewType";
import { Wishlist } from "./wishlistType";

export interface Product {
  id: string;
  productTitle: string;
  description: string;
  price: number;
  afterDiscountPrice: number;
  discountPercent?: number;
  discountPrice: number;
  couponPrice?: number;
  styleCode: string;
  sku: string;
  bottom?: string;
  comment: string;
  dupatta?: string;
  size: string;
  theme: string;
  top?: string;
  washcare?: string;
  length?: string;
  blouse?: string;
  type: string;
  isActive: boolean;
  isbannerImage: boolean;
  featuredProduct: boolean;
  isBestSeller: boolean;
  isDailyWear: boolean;
  isFestiveWear: boolean;
  isOccasionWear: boolean;
  thumbnail: string;
  bannerImage?: string;
  sizeChart?: string;
  sandokId?: string;
  categoriesId: string;
  tagId?: string;
  collectionId: string;
  productImage: string[];
  createdAt: Date;
  updatedAt: Date;
  MyCart: MyCartType[];
  categories: Categories;
  collection: Collection;
  ProductStock: ProductStock[];
  reviewAndRating: ProductRatingType[];
  // sandokProduct: SandokProduct[];
  Wishlist: Wishlist[];
  // NewOrderitem: NewOrderitem[];
  // Tag?: Tag;
  // MyOrder: MyOrder[];
}